+++ src/components/Footer/Footer.js
@@ -1,0 +1,21 @@
+import React from 'react';
+import './Footer.css';
+
+
+const Footer = () => {
+    return (
+        <div className='Box'>
+            <div className='Box1'>
+                IndoKhaadyam
+            </div>
+            <div className='triangle-div'>
+                <div className='Triangle'></div>
+            </div>
+            <div className='Box2'>
+                Made with 💖 in India for the world
+            </div>
+        </div>
+    )
+}
+
+export default Footer;
