+++ src/components/TopCategory/TopCategoryData.js
@@ -1,9 +1,9 @@
-import mithai from '../assets/mithai.jpg';
- import nasta from "../assets/nasta.jpg";
- import snacks from '../assets/snacks.jpg';
- import fastingfood from '../assets/fastingfood.jpg';
- import drink from '../assets/drink.jpg';
- import maincorse from '../assets/maincorse.jpg';
+import mithai from '../../assets/mithai.jpg';
+ import nasta from "../../assets/nasta.jpg";
+ import snacks from '../../assets/snacks.jpg';
+ import fastingfood from '../../assets/fastingfood.jpg';
+ import drink from '../../assets/drink.jpg';
+ import maincorse from '../../assets/maincorse.jpg';
 
 
  const data = [{
