+++ src/pages/Home/HomePage.jsx
@@ -1,11 +1,11 @@
 import React from "react";
-import Home from "../components/Home";
-import GlobalStyle from "../globalStyles";
-import Footer from "../components/Footer";
+import Home from "./Home";
+import GlobalStyle from "../../globalStyles";
+import Footer from "../../components/Footer/Footer";
 import { useMediaQuery } from "react-responsive";
-import Navbar from "../components/Navbar/Navbar";
-import MobileNavbar from "../components/Navbar/mobilenavbar";
-import TopCategory from "../components/TopCategory";
+import Navbar from "../../components/Navbar/Navbar";
+import MobileNavbar from "../../components/Navbar/mobilenavbar";
+import TopCategory from "../../components/TopCategory/TopCategory";
 
 
 const HomePage = () => {
